+++ packages/form/src/checkbox-group.vue
@@ -1,6 +1,14 @@
 <template>
   <div class="weui-cells_checkbox">
     <slot></slot>
+    <a
+      href="javascript:"
+      class="weui-cell weui-cell_active weui-cell_link"
+      v-if="showMore"
+      @click="more()"
+    >
+      <div class="weui-cell__bd">{{ moreText }}</div>
+    </a>
   </div>
 </template>
 <script>
@@ -15,6 +23,19 @@ export default {
       type: String,
       default: null,
     },
+    showMore: {
+      type: Boolean,
+      default: false
+    },
+    moreText: {
+      type: String,
+      default: '添加更多'
+    }
+  },
+  methods: {
+    more: function () {
+      this.$emit('onMore')
+    }
   }
 };
 </script>