+++ packages/form/src/control.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-form__control-area">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiFormControl'
+}
+</script>