+++ packages/form/src/radio.vue
@@ -1,12 +1,19 @@
 <template>
-  <label @click.prevent="onClickLabel" class="weui-cell weui-cell_active weui-check__label">
+  <label
+    @click.prevent="onClickLabel"
+    class="weui-cell weui-cell_active weui-check__label"
+  >
     <div class="weui-cell__bd">
-      <p>
-        <slot></slot>
-      </p>
+      <p>{{ label }}</p>
     </div>
     <div class="weui-cell__ft">
-      <input :name="name" :value="value" class="weui-check" type="radio" v-model="currentValue" />
+      <input
+        :name="name"
+        :value="value"
+        class="weui-check"
+        type="radio"
+        v-model="currentValue"
+      />
       <span class="weui-icon-checked"></span>
     </div>
   </label>
@@ -24,6 +31,7 @@ export default {
       type: String,
       default: null,
     },
+    label: String
   },
   computed: {
     currentValue: {