+++ packages/form/src/tip.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="weui-form__tips-area">
-    <p class="weui-form__tips" v-if="content" v-html="content"></p>
+    <p class="weui-form__tips" v-if="tipContent">{{ tipContent }}</p>
+    <p class="weui-form__tips" v-else-if="content">{{ content }}</p>
     <slot name="content"></slot>
   </div>
 </template>
@@ -8,6 +9,10 @@
 export default {
   name: 'WeuiFormTip',
   props: {
+    tipContent: {
+      type: String,
+      default: null
+    },
     content: {
       type: String,
       default: null